requirements.txt
aniso8601==9.0.1
blinker==1.7.0
cached-property==1.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-RESTful==0.3.10
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pytz==2025.2
requests==2.32.4
six==1.17.0
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
